requirements.txt
functions-framework>=3.0.0
requests>=2.20.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage==2.10.0
